+++ torchegranate/distributions/exponential.py
@@ -72,7 +72,8 @@ class Exponential(Distribution):
 			The dimensionality the distribution is being initialized to.
 		"""
 
-		self.scales = _cast_as_parameter(torch.zeros(d, device=self.device))
+		self.scales = _cast_as_parameter(torch.zeros(d, dtype=self.dtype,
+			device=self.device))
 
 		self._initialized = True
 		super()._initialize(d)
