+++ torchegranate/distributions/poisson.py
@@ -71,7 +71,8 @@ class Poisson(Distribution):
 			The dimensionality the distribution is being initialized to.
 		"""
 
-		self.lambdas = _cast_as_parameter(torch.zeros(d, device=self.device))
+		self.lambdas = _cast_as_parameter(torch.zeros(d, dtype=self.dtype,
+			device=self.device))
 
 		self._initialized = True
 		super()._initialize(d)
